DKG: more cleanups (#762)
* 1. fix a possible panic in case a message with larger than MAX_U16 polynomial is sent
2. add more comments
3. change Nodes::reduce to new_reduced so it can only be applied once

+++ fastcrypto-tbls/src/tests/polynomial_tests.rs
@@ -21,8 +21,8 @@ mod scalar_tests {
 
     #[test]
     fn test_degree<S: Scalar>() {
-        let s: u16 = 5;
-        let p = Poly::<S>::rand(s, &mut thread_rng());
+        let s: usize = 5;
+        let p = Poly::<S>::rand(s as u16, &mut thread_rng());
         assert_eq!(p.degree(), s);
     }
 
@@ -88,6 +88,7 @@ mod scalar_tests {
 mod points_tests {
     use super::*;
     use itertools::Either;
+
     #[test]
     fn test_eval_and_commit<G: GroupElement>() {
         // test zero
